migrate resetPasswordInitiate to /authManagement and add req res types
